battery-status/battery-full-01-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Status Test: BatteryManager - charging (full)</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/battery-status/">
<meta name="flags" content="interact">
<meta name="assert" content="The charging attribute must be set to true if the battery is full (the 'otherwise' condition).">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <h2>Preconditions</h2>
  <ol>
    <li>
      The device is plugged in to the charger before this test is run.
    </li>
    <li>
      The battery is full.
    </li>
  </ol>
</body>
<script>

async_test(function (t) {
  navigator.getBattery().then(function (battery) {
      t.step(function () {
        assert_true(battery.charging);
      });
      t.done();
    }, function (error) {
      t.step(function () {
        assert_unreached(error.message);
      });
      t.done();
    }
  )
}, document.title);

</script>
